This changeset implements goal-driven lazy m-Dijkstra/A*: after discovering each solution it backtracks and marks in CLOSED all nodes in the current path as been part of a solution path. When rewinding OPEN, only nodes flagged as been part of an optimal solution are re-expanded.
However, it can not work properly. An implementation of goal-driven
lazy m-Dijkstra/A* over mazes proves that it skips some solutions
(producing others of larger cost). The reason is simple: while it is
certainly true that new solution paths shall contain nodes that have
been previously discovered to be part of another solution path (as
disjoint solution paths are naturally unrevealed by dijkstra/A*), it
is not true that these nodes shall be precisely those in OPEN. A
proper implementation of the idea would require to re-expand nodes
whose path contains at least one node re-expanded but this is
worthless as shown below ---unless an alternative solution to this
problem is devised.

A simple example that proves the current version of goal-driven lazy
m-Dijkstra to be wrong is the following:

   t
 / |
s--A

Consider all edges have cost one. When expanding the nodes in OPEN
with g=1, say t is expanded first (throught the path <s, t>), then s
and t are flagged as part of a previous solution, so that after
rewinding OPEN, A would be skipped as it is not flagged.

The solution might be to re-expand nodes whose path contains at least
one that has been previously discovered to be part of a solution
path. In the exxample, this node would be s, but:

1. First, it makes no sense to mark s as s shall never be re-expanded.
2. Worst, all paths start at s, so that applying this rule would lead
to expand precisely the same nodes (the same number of times) of the
basic implementation of lazy m-Dijkstra/A*.