Implementation of labelednode_t
labelednode_t implements labeled backward nodes. These nodes are much
the same like backnodes, but there are a couple of key differences:

1. labelednode_t extends the definition of backpointer_t with the cost
of the operator that generated the child from its parent. The new
definition is called labeledbackpointer_t

2. Every instance of labelednode_t contains a vector of backward
g-values with the distinct values of all paths getting to t from that
node. They are stored separately in a vector of ints

labelednode_t provides means for implementing BELA*