Preparation of unit tests for BELA*
The class implementing BELA* is provided in this text though its
implementation is barely started. To ease its implementation, unit tests
are being developed for testing small parts of it. This changeset
contains a procedure for populating a closed list in a simple
domain (simplegrid_t) which shall be used in the development of BELA*