get_prefixes started
get_prefixes is an essential part of the implementation of BELA*. Among
other duties, it has to compute all optimal paths getting to the start
vertex of a centroid.

This changeset contains a primary implementation of get_prefixes in
BELA* which is tested with various unit test cases.

Still, get_prefixes has to discover centroids and to propagate backward
g-values ...