Implementation of get_paths
Every centroid is a representative of a class of paths. `get_paths`
computes the entire class given a centroid.

This changeset also contains a couple of unit tests for verifying its
correctness with the simple grid or the genreal grid. Nevertheless, it
is still necessary to test it with true centroids in the genral grid