Last unit test for testing get_paths
This test verifies that the computation of all paths represented by true
centroids (i.e., centroids representing suboptimal paths that are
discovered once the optimal paths have been revealed) actually works.

This test actually showed a flaw in the computation of the prefixes: the
goal can never be part of any prefix because we never want the goal
state to be expanded more than once in a single path. However, when
considering the last edge to get to the goal through an optimal path as
a centroid, the computation of prefixes considered the goal state to be
part of it. As a matter of fact, there should be never a centroid
arising from the goal state ...