Bug fixed in the definition of goal in grid_t
Certainly, goals must never create new centroids. In my last unit test I
took specific care about this in the computation of get_prefixes but
this is absolutely unnecessary. The reason is that the goal should never
be expanded so that no node would have a backpointer to it and thus, it
would never be possible to create a centroid from it.

It turned out there was a bug in the definition of the goal in the
definition of the grid domain. Once fixed everything worked out
perfectly. Other modifications were necessary in other unit tests as a
result of this modification