Implementation of BELA*
The implementation of BELA* is almost finished. There is one case to
deal with though. The open list might eventually become empty and still
there are new non-simple paths to found. This case is not implemented
yet.

This changeset contains a couple of additional unit tests to verify that
BELA* is able to find either one single optimal solution or two best
solutions. When k>1 BELA* is shown to be clearly superior to mA*