Preliminary implementation of a generic solver
This changeset provides a preliminary implementation of a generic solver
for the N-pancake domain. Still, there are a good number of features to
add but in the meantime, this version can be used to start any solver
over a selection of pancake instances.

Results show an outstanding difference in performance wrt mA*