Flag no_doctor added
It is feasible to compute a huge number of distinct paths In the square
grid domain (with no obstacles). However, the results take a very long
time to appear because the doctor takes a quadratic time (on the number
of solution paths) to verify there are no duplicates. Hence, a new flag,
--no-doctor, has been added which disables the automated verification of
the solutions