Bound on the number of prefixes/suffixes
A significant improvement in runtime of BELA* is possible by bounding
the number of prefixes/suffixes to compute. In some cases, there might
be a lot of preffixes/suffixes so that instead of awaiting for them all
to be computed and then to prune the construction of solution paths when
K is achieved, it is much better to produce only the minimal amount of
prefixes/suffixes. Other strategies are certainly possible ...