Summary option added to all solvers
In the end, the information of every solution path is almost never used
in so many experiments. Thus, a new flag has been added to all solvers:
'--summary' instructs the khs library to provide information only about
the last solution path (i.e., about the effort to compute the k shortest
paths). The resulting csv files are far more manageable in particular
for large values of k.

Still, detailed information is available. For this, '--summary' should
not be given and the information of every single solution path is then
produced ---which results in very large .csv files