New version of the plotter
The option '-c' has been removed. In addition, get_data was not working
with .xlsx files and this has been fixed. Finally, the package csv is
not necessary at all as pyexcel can read any type of spreadsheets and
thus it is preferred to use the same (simple) interface to process any
spreadsheet