Tiny modification in the creation of new centroids
In the brute-force variant of BELA*, centroids are checked first by
verifying that the g*-value of the node to expand (after verifying it is
not the goal and also, that it has been previously expanded) is smaller
than the g-value of the new path. This is unnecessary: the next step is
to create a centroid for every gb-value but gb-values can only exist if
its g*-value is smaller than the g-value of the new path in the
brute-force variant.

Also, this is exactly how it is done in the heuristic variant of the
algorithm, so that the pseudocode of both variants is now the same